+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Add `expectedItemsCount` _optional_ argument to `parse` method of `ITuple` to allow explicit validation of parsed result
 - Require higher version (`^2.9.3`) of `beberlei/assert` library because of simplified error messages
 - Allow `beberlei/assert` library in version `^3.0`
+- Add `parseMatch` and `parseMatchTypes` to `ITuple` to validate parsed result
+- Simplify `toString` method of `Tuple`
 
 ## 3.5.1 - 2018-07-13
 - Fix `implode` on `ListCollection`